DeviceKornet Начало проекта Visual Studio

+++ sources/Kornet/Device/Utils/Math.h
@@ -26,7 +26,7 @@ public:
         }
         else
         {
-            value = min;
+            value = (uint8)min;
         }
     }
